requirements.txt
appnope==0.1.2
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
Bottleneck==1.3.4
brotlipy==0.7.0
CacheControl==0.12.10
cachy==0.3.0
certifi==2022.12.7
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.12
cleo==0.8.1
click~=8.1.3
clikit==0.6.2
cloudpickle==2.0.0
colorama==0.4.4
conda==4.12.0
conda-package-handling==1.8.1
coverage==6.3.2
crashtest==0.3.1
cryptography==36.0.0
cvxopt==1.3.0
cycler==0.11.0
decorator==5.1.1
deprecation==2.1.0
DiffResBP-Simulator==0.1.0
distlib==0.3.4
executing==0.8.3
filelock==3.6.0
fonttools==4.31.2
future==0.18.2
graphviz==0.19.1
html5lib==1.1
hyperopt~=0.2.7
identify==2.4.12
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
intervaltree==3.1.0
invoke~=1.7.3
ipython==8.2.0
jedi==0.18.1
jellyfish==0.9.0
Jinja2==3.1.1
jsonpickle==2.1.0
keyring==23.5.0
kiwisolver==1.4.2
lockfile==0.12.2
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib~=3.6.0
matplotlib-inline==0.1.3
mpmath==1.2.1
msgpack==1.0.3
munkres==1.1.4
networkx~=2.8.7
nodeenv==1.6.0
numexpr==2.8.1
numpy~=1.23.4
packaging==21.3
pandas~=1.5.2
parso==0.8.3
pastel==0.2.1
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pip==21.2.4
pkginfo==1.8.2
platformdirs==2.5.1
pluggy==1.0.0
pm4py==2.2.19.2
pm4py-wrapper==1.1.4
poetry==1.1.13
poetry-core==1.0.8
pre-commit==2.18.1
prompt-toolkit==3.0.28
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycosat==0.6.3
pycparser==2.21
pydotplus==2.0.2
Pygments==2.11.2
pylev==1.4.0
pymongo==3.12.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
pytest~=7.1.3
pytest-cov==3.0.0
python-dateutil==2.8.2
pytz==2022.1
pytzdata==2020.1
pyvis==0.1.9
PyYAML==6.0
requests==2.27.1
requests-toolbelt==0.9.1
ruamel-yaml-conda==0.15.100
scipy~=1.9.2
setuptools~=63.4.3
shellingham==1.4.0
simod==3.1.0
simpy==4.0.1
six==1.16.0
sortedcontainers==2.4.0
stack-data==0.2.0
StringDist==1.0.9
sympy==1.10.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.10.1
tornado==6.1
tqdm~=4.64.1
traitlets==5.1.1
ukkonen==1.0.1
urllib3==1.26.9
virtualenv==20.14.1
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.37.1
xmltodict~=0.13.0
zipp==3.8.0
statistics~=1.0.3.5
fastapi~=0.89.1
pydantic~=1.10.3